Use hook to control fixture dynamically
This PR introduces adds a hook to bypass fixture according to markers.
The hook checks asic marker of each test item, and determines whether
fixture check_dut_asic_type is added.
